nushell LSP and better t function

+++ nushell/templates/script.nu
@@ -1,6 +1,21 @@
 def t [t: path] {
     let type = $t | path parse | get extension
     let template: path = "~/.config/nushell/templates/template." + $type | path expand
+
+    # File exist
+    if ($t | path expand | path exists) {
+        match $type {
+            "typ" => {
+                typst compile $t
+                let pdf = ($t | path parse | get stem) + (".pdf")
+                start $pdf
+            },
+        }
+        nvim $t
+        return
+    }
+
+    # File does not
     if ($template | path exists) {
         cp $template $t
         match $type { 
@@ -10,7 +25,8 @@ def t [t: path] {
                 start $pdf
             },
         }
-        return $t
+        nvim $t
+        return
     } else {
         print $"no template for filetype: '($type)'"
         touch $t
